+++ packages/mta/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.2](https://github.com/hiddentao/camomail/compare/v1.2.1...v1.2.2) (2020-04-13)
+
+**Note:** Version bump only for package @camomail/mta
+
+
+
+
+
+## [1.2.1](https://github.com/hiddentao/camomail/compare/v1.2.0...v1.2.1) (2020-04-13)
+
+**Note:** Version bump only for package @camomail/mta
+
+
+
+
+
 # [1.2.0](https://github.com/hiddentao/camomail/compare/v1.1.8...v1.2.0) (2020-04-12)